HIS-83: implement sidebar
