--- histree-frontend/src/components/comparison/ComparisonToggle.scss
@@ -1,5 +1,0 @@
-.comparison_toggle_container {
-	margin: 1.5em 0.5em 0 0;
-	width: 5%;
-	height: 5%;
-}
