+++ histree-frontend/src/components/comparison/ModeIcons.scss
@@ -1,0 +1,14 @@
+.comparison_toggle_container {
+	display: flex;
+	flex-direction: column;
+	padding: 0.5em;
+	margin: 1em 0 0 0;
+	background-color: white;
+	border-radius: 0.3em;
+	border: 1px solid lightgray;
+}
+
+.sidebar_icon_button {
+	height: 100%;
+	padding-bottom: 2em;
+}
