+++ histree-frontend/src/components/comparison/index.ts
@@ -1,3 +1,3 @@
 export * from "./ComparisonCard";
-export * from "./ComparisonToggle";
+export * from "./ModeIcons";
 export * from "./ComparisonFlow";
