+++ histree-frontend/src/pages/TreePage.scss
@@ -18,4 +18,5 @@
 	left: 2%;
 	top: 2%;
 	display: flex;
+	align-items: flex-start;
 }
